# docs/swagger.yaml
basePath: /api/v1
definitions:
  auth.LoginInput:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  auth.LoginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/auth.UserResponse'
    type: object
  auth.RegisterOwnerInput:
    properties:
      business_name:
        type: string
      email:
        type: string
      owner_name:
        type: string
      password:
        type: string
    type: object
  auth.UserResponse:
    properties:
      branch_id:
        type: string
      business_id:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      role:
        type: string
    type: object
  branch.Branch:
    properties:
      address:
        type: string
      business_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
    type: object
  branch.CreateBranchInput:
    properties:
      address:
        type: string
      name:
        type: string
    type: object
  branch.UpdateBranchInput:
    properties:
      address:
        type: string
      is_active:
        type: boolean
      name:
        type: string
    type: object
  business.Business:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      service_charge_percent:
        type: number
      slug:
        type: string
      tax_percent:
        type: number
    type: object
  business.UpdateBusinessInput:
    properties:
      name:
        type: string
      service_charge_percent:
        type: number
      tax_percent:
        type: number
    type: object
  dashboard.BranchRev:
    properties:
      branch_id:
        type: string
      branch_name:
        type: string
      revenue:
        type: number
    type: object
  dashboard.CashierDashboard:
    properties:
      open_orders_count:
        type: integer
      orders_handled_today:
        type: integer
      total_collected_today:
        type: number
    type: object
  dashboard.OwnerDashboard:
    properties:
      revenue_per_branch:
        items:
          $ref: '#/definitions/dashboard.BranchRev'
        type: array
      top_menu_items:
        items:
          $ref: '#/definitions/dashboard.TopMenuItem'
        type: array
      total_orders_today:
        type: integer
      total_revenue_today:
        type: number
    type: object
  dashboard.TopMenuItem:
    properties:
      menu_item_id:
        type: string
      menu_item_name:
        type: string
      total_sold:
        type: integer
    type: object
  httputil.Error400Response:
    properties:
      error:
        example: invalid request body
        type: string
      success:
        example: false
        type: boolean
    type: object
  httputil.Error401Response:
    properties:
      error:
        example: unauthorized
        type: string
      success:
        example: false
        type: boolean
    type: object
  httputil.Error403Response:
    properties:
      error:
        example: forbidden accesss
        type: string
      success:
        example: false
        type: boolean
    type: object
  httputil.Error404Response:
    properties:
      error:
        example: resource not found
        type: string
      success:
        example: false
        type: boolean
    type: object
  httputil.Error500Response:
    properties:
      error:
        example: internal server error
        type: string
      success:
        example: false
        type: boolean
    type: object
  httputil.MessageResponse:
    properties:
      data:
        example: ok
        type: string
      success:
        example: true
        type: boolean
    type: object
  menu.CreateMenuItemInput:
    properties:
      category:
        type: string
      description:
        type: string
      is_available:
        type: boolean
      modifier_group_ids:
        items:
          type: string
        type: array
      name:
        type: string
      price:
        type: number
    type: object
  menu.MenuItemResponse:
    properties:
      business_id:
        type: string
      category:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_available:
        type: boolean
      modifiers:
        items:
          $ref: '#/definitions/menu.ModifierGroupResponse'
        type: array
      name:
        type: string
      price:
        type: number
    type: object
  menu.ModifierGroupResponse:
    properties:
      id:
        type: string
      is_required:
        type: boolean
      max_select:
        type: integer
      name:
        type: string
      options:
        items:
          $ref: '#/definitions/menu.ModifierOptionResponse'
        type: array
    type: object
  menu.ModifierOptionResponse:
    properties:
      extra_price:
        type: number
      group_id:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  menu.UpdateMenuItemInput:
    properties:
      category:
        type: string
      description:
        type: string
      is_available:
        type: boolean
      modifier_group_ids:
        items:
          type: string
        type: array
      name:
        type: string
      price:
        type: number
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Multi-tenant POS SaaS backend
  title: Bayar.in API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return a JWT token
      parameters:
      - description: Login payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.Error401Response'
      summary: Login
      tags:
      - Auth
  /auth/logout:
    post:
      description: Revoke the current session token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httputil.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.Error500Response'
      security:
      - BearerAuth: []
      summary: Logout
      tags:
      - Auth
  /auth/me:
    get:
      description: Returns the authenticated user's profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.UserResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.Error404Response'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - Auth
  /auth/register-owner:
    post:
      consumes:
      - application/json
      description: Creates a new business and an owner account atomically
      parameters:
      - description: Registration payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterOwnerInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/auth.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Error400Response'
      summary: Register a new business owner
      tags:
      - Auth
  /branches:
    get:
      description: Returns all branches that belong to the authenticated owner's business
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/branch.Branch'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.Error500Response'
      security:
      - BearerAuth: []
      summary: List branches
      tags:
      - Branches
    post:
      consumes:
      - application/json
      description: Create a new branch under the authenticated owner's business
      parameters:
      - description: Branch payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/branch.CreateBranchInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/branch.Branch'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Error400Response'
      security:
      - BearerAuth: []
      summary: Create branch
      tags:
      - Branches
  /branches/{id}:
    patch:
      consumes:
      - application/json
      description: Partially update a branch's name, address, or active status
      parameters:
      - description: Branch UUID
        in: path
        name: id
        required: true
        type: string
      - description: Update payload (all fields optional)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/branch.UpdateBranchInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/branch.Branch'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Error400Response'
      security:
      - BearerAuth: []
      summary: Update branch
      tags:
      - Branches
  /business:
    get:
      description: Returns the authenticated owner's business profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/business.Business'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.Error404Response'
      security:
      - BearerAuth: []
      summary: Get business
      tags:
      - Business
    patch:
      consumes:
      - application/json
      description: Partially update business name, tax percent, or service charge
        percent
      parameters:
      - description: Update payload (all fields optional)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/business.UpdateBusinessInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/business.Business'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Error400Response'
      security:
      - BearerAuth: []
      summary: Update business
      tags:
      - Business
  /dashboard/cashier:
    get:
      description: 'Returns cashier-level stats: today''s transactions, open orders,
        and revenue for this session'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dashboard.CashierDashboard'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.Error500Response'
      security:
      - BearerAuth: []
      summary: Cashier dashboard
      tags:
      - Dashboard
  /dashboard/owner:
    get:
      description: 'Returns business-level stats: total revenue, orders, top menu
        items, and per-branch breakdown'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dashboard.OwnerDashboard'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.Error500Response'
      security:
      - BearerAuth: []
      summary: Owner dashboard
      tags:
      - Dashboard
  /menu:
    get:
      description: List all menu items with nested modifier groups and options
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/menu.MenuItemResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.Error401Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httputil.Error403Response'
      security:
      - BearerAuth: []
      summary: List menu items
      tags:
      - menu
    post:
      consumes:
      - application/json
      description: Create a new menu item and optionally link modifier groups
      parameters:
      - description: Menu item payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/menu.CreateMenuItemInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/menu.MenuItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Error400Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.Error401Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httputil.Error403Response'
      security:
      - BearerAuth: []
      summary: Create menu item
      tags:
      - menu
  /menu/{id}:
    patch:
      consumes:
      - application/json
      description: Update menu item fields and replace modifier group links
      parameters:
      - description: Menu item UUID
        in: path
        name: id
        required: true
        type: string
      - description: Update payload — modifier_group_ids replaces all existing links
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/menu.UpdateMenuItemInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/menu.MenuItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Error400Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.Error401Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httputil.Error403Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.Error404Response'
      security:
      - BearerAuth: []
      summary: Update menu item
      tags:
      - menu
...
securityDefinitions:
  BearerAuth:
    description: 'Paste your JWT token as: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"